Started breaking up REPL
Began refactoring of the REPL to make other processing easier.
