+++ Elision/src/ornl/elision/repl/Repl.scala
@@ -47,71 +47,6 @@ import java.io.{FileWriter, FileReader, BufferedReader}
 import ornl.elision.core.OperatorLibrary
 
 /**
- * Provide information about the current version of Elision.  This information
- * is obtained from the `configuration.xml` file expected to be in the root
- * of the current classpath (thus in the jar file).  It has the following
- * format.
- * 
- * {{{
- * <configuration
- *   name = "name of program"
- *   maintainer = "name and email of maintainer"
- *   web = "web address of program">
- *   <version
- *     major = "major version number"
- *     minor = "minor version number"
- *     build = "build date / time identifier"/>
- * </configuration>
- * }}}
- * 
- * The `loaded` field reveals if the file was successfully loaded.  If not, the
- * other fields contain default (and useless) information.  Missing fields
- * result in a value of the empty string for that field.  Extra fields not
- * specified above are ignored, but may be used in the future.
- */
-object Version {
-  /** Name of the program. */
-  var name = "Elision"
-    
-  /** Name and email of the maintainer. */
-  var maintainer = "<maintainer>"
-    
-  /** Web address for the program. */
-  var web = "<web>"
-    
-  /** Major version number (integer string). */
-  var major = "UNK"
-    
-  /** Minor version number (integer string). */
-  var minor = "UNK"
-    
-  /** Twelve digit build identifier.  Date and time. */
-  var build = "UNK"
-    
-  /** If true then data was loaded.  If false, it was not. */
-  var loaded = false
-
-  private def init {
-	  // Open the file.  We expect to find config.xml in the classpath.
-	  val config_resource = getClass.getResource("/configuration.xml")
-	  if (config_resource != null) {
-	    // Parse the file.
-	    val config = scala.xml.XML.load(config_resource.toString())
-	    
-	    // Pull out the data and override the local defaults.
-	    name = config.\("@name").text
-	    maintainer = config.\("@maintainer").text
-	    web = config.\("@web").text
-	    major = config.\("version").\("@major").text
-	    minor = config.\("version").\("@minor").text
-	    build = config.\("version").\("@build").text
-	    loaded = true
-	  }
-  }
-  init
-}
-
-/**
  * Provide a REPL to experiment with the new term rewriter.
  * 
  * ==Use==
